% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelations.R
\name{calcStats}
\alias{calcStats}
\title{Calculates correlations for the test case. The test case is designed to
calculate the data correctness in the data frame generated by
computeCorrelations(). This helper counts the records in the
data frame that belong to bioSys and sums-up the correlation and
semantic simulation columns.
Those values are later compared in the test cases with pre-calculated values}
\usage{
calcStats(dframe, bioSys)
}
\arguments{
\item{dframe}{- the source data frame}

\item{bioSys}{- name of the system (PHALY, SLIGR or NLRIN)}
}
\value{
(list) ## which contains record counts, control sum of the correlation
and sem sim correlation for all records that belong to a System
in the data frame
}
\description{
Calculates correlations for the test case. The test case is designed to
calculate the data correctness in the data frame generated by
computeCorrelations(). This helper counts the records in the
data frame that belong to bioSys and sums-up the correlation and
semantic simulation columns.
Those values are later compared in the test cases with pre-calculated values
}
\author{
\href{https://orcid.org/0000-0002-8778-6442}{Denitsa Vasileva} (aut)
}
