% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_distance_matrices.R
\name{combine_distance_matrices}
\alias{combine_distance_matrices}
\title{Combine Distance Matrices}
\usage{
combine_distance_matrices(mode, distanceMatrices)
}
\arguments{
\item{mode}{Method by which to combine distance matrices}

\item{distanceMatrices}{List of objects of class distance.}
}
\value{
single object of class distance.
}
\description{
\code{combine_distance_matrices} This is a helper function for \code{\link{clusterSystems}} and is unlikely to be useful elsewhere.
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
