% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunitcoExpressionAnalysis.R
\name{complexExpHeatmap}
\alias{complexExpHeatmap}
\title{\code{complexExpHeatmap} Heatmap of the complex subunits expression profile}
\usage{
complexExpHeatmap(complex)
}
\arguments{
\item{complex}{(char)  name of the complex}
}
\value{
(list) A plot representing expression heatmap of the complex subunits
                 across all microarray experiments
}
\description{
\code{complexExpHeatmap} Generate a Heatmap of complex subunits expression profile
for visualization. Each row for the heatmap displays the gene encoding for each
subunit in HGNC symbols, whereas each column displays an experiment. For each grid,
the more navy it is, the more negatively expressed the gene is. Vice versa for the
brightness of yellow representing positive expression
}
\examples{
\dontrun{
complex <- "MDC1-MRN-ATM-FANCD2 complex"
heatmap <- complexExpHeatmap(complex) # give the expression profile heatmap
                                      # for visualization
heatmap # display the heatmap
}
}
