% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelations.R
\name{computeCorrelations}
\alias{computeCorrelations}
\title{\code{computeCorrelations()} Calculates pairwise Semantic similarity and co-expression correlation
between expression profiles of genes and returns a
data frame}
\usage{
computeCorrelations(silent = FALSE)
}
\arguments{
\item{silent}{= FALSE suppresses all messages when TRUE}
}
\value{
(data frame)           containing gene A, gene B, correlation between them and
                               GOSemSim correlation
[GOSemSim package](http://bioconductor.org/packages/release/bioc/html/GOSemSim.html)
[org.Hs.eg.db data package](https://bioconductor.org/packages/release/data/annotation/html/org.Hs.eg.db.html)
}
\description{
\code{computeCorrelations()} Calculates pairwise Semantic similarity and co-expression correlation
between expression profiles of genes and returns a
data frame
}
\seealso{
\code{\link{<fetchData>}}
}
\author{
\href{https://orcid.org/0000-0002-8778-6442}{Denitsa Vasileva} (aut)
}
