% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGI_PPI_Integrator.R
\name{convertToHGNC}
\alias{convertToHGNC}
\title{Convert a system's components to HGNC-annotated physical interactors}
\usage{
convertToHGNC(myGenes, mart)
}
\arguments{
\item{myGenes}{A list of system's ENSEMBL peptide IDs}

\item{mart}{A biomaRt mart to be queried for ENSEMBL ID conversion}
}
\value{
The dataframe of physically interacting genes in \code{myGenes} according to \code{mart}
}
\description{
Convert a system's components to HGNC-annotated physical interactors
}
