% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprTrajectory.R
\name{createExprTrajectory}
\alias{createExprTrajectory}
\title{Title.}
\usage{
createExprTrajectory(filename)
}
\arguments{
\item{filename}{The filename of the .fcs flow cytometry file.}
}
\value{

}
\description{
\code{createExprTrajectory} Creates the p-Creode Graphs from flow cytometry
}
\details{
Details.
}
\section{Dependency}{
 the pcreode package, which can be installed with:
\code{sudo pip install pcreode}
}

\examples{
# Loads the Myeloid Differentiation Flow Cyt. Dataset, and stores the trajectory graphs
createExprTrajectory("./inst/extdata/rna_seq_myeloid.fcs")

}
\author{
Yoonsik Park (au)
}
