% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPlot.R
\name{myPlot}
\alias{myPlot}
\title{\code{myPlot()} Plots co-expression correlation vs semantic similarity and their own heatmaps respectively.}
\usage{
myPlot(geneSet)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{geneSet}{
  set of genes with HGNC symbols of interest
}

\value{
 \item{heatmap1 }{heatmap of expression correlation of each pair of genes in the geneSet}
 \item{heatmap2 }{heatmap of GO semantic similarity of each pair of genes in the geneSet}
 \item{lmplot }{ggplot of correlation vs semantic similarity}
 \item{result}{summary of linear model of correlation vs semantic similarity}
%% ...
}
\references{
http://www.sthda.com/english/wiki/ggcorrplot-visualization-of-a-correlation-matrix-using-ggplot2
https://www.bioconductor.org/packages/devel/bioc/vignettes/GOSemSim/inst/doc/GOSemSim.html
}
\author{
\herf{https://orcid.org/0000-0003-1275-7951}{Yuhan Zhang}(aut)
}
\description{
\code{myPlot()} Calculates pairwise co-expression Correlation
between expression profiles of genes in input geneSet and pairwise
Semantic Similarity between GO terms of input geneSet. Plot the heatmap of these correlations and similarities and then plot scatterplot and build lm model for correlation vs similarity
}
\examples{
\dontrun{
geneSet <- c("AMBRA1","ATG14","ATP2A1","ATP2A2","ATP2A3")
myPlot(geneSet)
## plots pairwise expression correlation heatmap graph, semantic
## similarity heatmap  and correlation vs similarity
## graphs 
}