% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayEnrich.R
\name{EnrichmentHelper}
\alias{EnrichmentHelper}
\title{\code{EnrichmentHelper} It's a helper function for pathway enrichment analysis 
function load all the required data for pathway enrichment analysis
Return a list that contain reactome information. Modified data annotation from 
reference: https://github.com/yoonsikp/BCB420.2019.REACTOME/}
\usage{
EnrichmentHelper()
}
\value{
A list of annotated information for r eactome
}
\description{
\code{EnrichmentHelper} It's a helper function for pathway enrichment analysis 
function load all the required data for pathway enrichment analysis
Return a list that contain reactome information. Modified data annotation from 
reference: https://github.com/yoonsikp/BCB420.2019.REACTOME/
}
\examples{
tmp<-EnrichmentHelper()

}
\author{
{Yufei Yang} (aut)
}
