% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_dist.R
\name{expr_dist}
\alias{expr_dist}
\title{Expression Distance.}
\usage{
expr_dist(gene1, gene2, profiles)
}
\arguments{
\item{gene1}{Character, HGNC symbol of the first gene.}

\item{gene2}{Character, HGNC sysmbol of the second gene}

\item{profiles}{Matrix of expression profiles, generated by calling
fetchData("GEOprofiles")}
}
\value{
Numeric, distance score: the absolute value of the spearman
correlation of the expression profiles of gene1 and gene2. Distance is
returned as 1 if the expression profile for either of the genes is unavailable.
}
\description{
\code{expr_dist} calculates the pairwise distance of gene expression profiles.
}
\details{
distance between 2 expression profiles is calculated by taking the absolute
value of the Spearman correlation. Spearman correlation is used because we do
not care whether the correlation is linear. Absolute value ensures that both strong
positive and negative correlations result in high distance.
}
\examples{
# Find the distance in the expression profiles of BRCA1 and BRCA2
GEO <- fetchData("GEOprofiles")
expr_dist("BRCA1", "BRCA2", GEO)


}
\seealso{
\code{\link{fetchData}} for the format of expression profiles used by
this function.
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
