% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_sim.R
\name{expr_sim}
\alias{expr_sim}
\title{Expression Similarity.}
\usage{
expr_sim(gene1, gene2, profiles = exprofs)
}
\arguments{
\item{gene1}{Character, HGNC symbol of the first gene.}

\item{gene2}{Character, HGNC sysmbol of the second gene}

\item{profiles}{Matrix of expression profiles, generated by loading "http://steipe.biochemistry.utoronto.ca/abc/assets/GEO-QN-profile-2019-03-24.rds"}
}
\value{
Numeric, similarity score: the absolute value of the spearman correlation of the expression profiles of gene1 and gene2. Or NA if the expression profile for that gene is unavailable.
}
\description{
\code{expr_sim} Calculate the pairwise similarity of gene expression profiles.
}
\details{
Similarity between 2 expression profiles is calculated by taking the absolute value of the Sprearman correlation. Spearman correlation is used because we do not care whether the correlation is linear. Absolute value ensures that both strong positive and negative correlations result in high similarity.
}
\section{<title>}{
 additional explanation
}

\examples{
# Find the similarity in the expression profiles of BRCA1 and FANCI
expr_sim("BRCA1", "FANCI")
# [1] 0.2387499

}
\seealso{
\code{\link{<function>}} <describe related function>, ... .

Other <optional description of family>: \code{\link{make_matrix}},
  \code{\link{net_sim}}, \code{\link{tf_sim}}
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
\concept{<optional description of family>}
