% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exProf.R
\name{exProf}
\alias{exProf}
\title{Gene Expression Profile.}
\usage{
exProf(sym, myQNXP = myQNXP)
}
\arguments{
\item{sym}{(character) Symbol of gene being investigated}

\item{myQNXP}{(matrix) Matrix of expressions. Columns represent experiments and
rows individual genes}
}
\value{
(numeric) Vector of log-normalize expressions by condition
}
\description{
\code{exProf} Returns the log-normalized expression values for gene provided.
}
\author{
\href{https://orcid.org/0000-0002-2821-6204}{Matthew McNei;} (aut)
}
