% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSub.R
\name{fectchComponentNotInSub}
\alias{fectchComponentNotInSub}
\title{\code{fectchComponentNotInSub} Find components that don't appear in any subsystem
                               in the provided system .}
\usage{
fectchComponentNotInSub(systemName, db)
}
\arguments{
\item{systemName}{(character) The system name that have 6 characters. All characters}

\item{db}{(character) The database.}
}
\value{
(character) A list of components in the system but not in the subsystem
}
\description{
\code{fectchComponentNotInSub} Find components that don't appear in any subsystem
                               in the provided system .
}
\examples{
mySDB <- fetchData("SysDB")
fectchComponentNotInSub("PHALLY", mySDB)
}
