% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchComponents.R
\name{fetchComponents}
\alias{fetchComponents}
\title{\code{fetchComponents} get all gene symbols in a system.}
\usage{
fetchComponents(sys)
}
\arguments{
\item{sys}{(character)  Name of a system}
}
\value{
(character) a vector of HGNC symbols
}
\description{
\code{fetchComponents} is currently only a stub.
}
\examples{
fetchComponents("PHALY")  # all PHALY system proteins
}
