% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchData.R
\name{fetchData}
\alias{fetchData}
\title{\code{fetchData} read a prepared dataset from a repository.}
\usage{
fetchData(set)
}
\arguments{
\item{set}{(character)  name of the requested dataset}
}
\value{
a dataset
}
\description{
\code{fetchData} documentation forthcoming.
}
\details{
For the BioGRID dataset, see details in the \href{../doc/dataDetails-BioGRID.html}{ BioGRID data vignette} (or load the vignette with \code{vignette("dataDetails-BioGRID", package = "BCB420.2019.ESA")}).
}
\examples{
fetchData()                         # prints available datasets
HGNC <- fetchData("HGNCreference")  # assigns the HGNC reference dataset
}
