% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchPDBHGNCdatabase.R
\name{fetchPDBHGNCdatabase}
\alias{fetchPDBHGNCdatabase}
\title{\code{fetchPDBHGNCdatabase} Retrieve the database containing
the annotated database for HGNC and PDB relations created by author
Heewon Lee from previous BCB420 course project in the
BCB420.2019.PDB package on GitHub linked here:
https://github.com/judyheewonlee/BCB420.2019.PDB}
\usage{
fetchPDBHGNCdatabase()
}
\value{
A dataframe contianing the PDB-HGNC annotated database
linking PDB ID to HGNC ID with annotations from the
BCB420.2019.PDB package. This is used in the SeqComparisonTable() function
in the BCB420.2019.ESA package.
}
\description{
\code{fetchPDBHGNCdatabase} Retrieve the database containing
the annotated database for HGNC and PDB relations created by author
Heewon Lee from previous BCB420 course project in the
BCB420.2019.PDB package on GitHub linked here:
https://github.com/judyheewonlee/BCB420.2019.PDB
}
\examples{

# Call the fetchPDBHGNCdatabase function to generate a dataframe
# containing the annotated PDB-HGNC database from the
# BCB420.2019.PDB package
pdbHGNCdata <- fetchPDBHGNCdatabase()

}
\author{
{Heewon Lee} (aut)
}
