% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSub.R
\name{fetchSubsystem}
\alias{fetchSubsystem}
\title{\code{fetchSubsystem} Find subsystem in the provided system.}
\usage{
fetchSubsystem(db, sys)
}
\arguments{
\item{db}{(character) The database.}

\item{systemName}{(character) The system name that have 6 characters. All characters}
}
\value{
(character) A list of subsystem in the system
}
\description{
\code{fetchSubsystem} Find subsystem in the provided system.
}
\examples{
mySDB <- fetchData("SysDB")
fetchSubsystem(mySDB, "PHALLY")

}
