% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSTRINGKnodes.R
\name{findSTRINGKnodes}
\alias{findSTRINGKnodes}
\title{\code{findSTRINGKnodes} output a the K node values for vertices in the STRING interaction graph given the system of genes 'sys'.}
\usage{
findSTRINGKnodes(sys, per = 0.05)
}
\arguments{
\item{sys}{(character)       5-character system code of genes}

\item{per}{(numeric)         Percentile of results to be plotted. Default is 0.05}
}
\value{
(numeric) Knode values
}
\description{
\code{findSTRINGKnodes} outputs a the K node values (as defined by Cornish & Markowetz, 2014),
a measure of the vertex association to the high-weighted vertices defined by the input
curated gene system 'sys'. Also plots the knode values for the top percentile of knode values
given by 'per'
}
\examples{
findSTRINGKnodes("PHALY", 0.05)
}
