% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSub.R
\name{findSub}
\alias{findSub}
\title{\code{findSub} Find new subsystem for the system we asked.}
\usage{
findSub(systemName, threshold, db)
}
\arguments{
\item{systemName}{(character) The system name that have 6 characters. All characters
are uppercase.e.g. PHALLY}

\item{threshold}{(integer) The thershold for selection possible subsystem.}

\item{db}{(character) The database.}
}
\value{
(table) a table with possible subsystem name and its components.
}
\description{
\code{findSub} Find new subsystem for the system we asked.
}
\examples{
findsub("PHALLY", 0.2)
}
