% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSub.R
\name{findSubInCom}
\alias{findSubInCom}
\title{\code{findSubInCom} Find subsystems in the provided components by using hu.map database.}
\usage{
findSubInCom(systemName, threshold, db)
}
\arguments{
\item{systemName}{(character) The system name that have 6 characters. All characters}

\item{threshold}{(integer) The thershold for selection possible subsystem}

\item{db}{(character) The database.}
}
\value{
(character) A list of components in the system but not in the subsystem
}
\description{
\code{findSubInCom} Find subsystems in the provided components by using hu.map database.
}
\examples{
mySDB <- fetchData("SysDB")
findSubIncom("PHALLY", mySDB)
}
