% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunitcoExpressionAnalysis.R
\name{genePairsShuffle}
\alias{genePairsShuffle}
\title{\code{genePairsShuffle} Generate Coexpression correlation generated by
shuffled expression values of the two genes}
\usage{
genePairsShuffle(gene1, gene2)
}
\arguments{
\item{gene1}{(char)  HGNC symbol for the first gene of the coexpression
profile}

\item{gene2}{(char)  HGNC symbol for the second gene}
}
\value{
(double vector) A double vector of Coexpression correlation generated
by shuffled expression values of the two genes
}
\description{
\code{genePairsShuffle} Coexpression correlation created by shuffled
expression values of the two genes by 100000 times
}
\examples{
\dontrun{
gene1 <- "NBN"
gene2 <- "ATM"
corrShuffled <- genePairsShuffle(gene1, gene2)
corrShuffled
}

}
