% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunitcoExpressionAnalysis.R
\name{getCoExpProfile}
\alias{getCoExpProfile}
\title{\code{getCoExpProfile} Generate a coexpression profile}
\usage{
getCoExpProfile(complex)
}
\arguments{
\item{complex}{(char)  name of the complex}
}
\value{
(data frame) a data frame with there columns: the first gene, the second gene
and the Pearson Correlation Coefficients for their coexpression
}
\description{
\code{getCoExpProfile} Generate a coexpression profile
}
\examples{
complex <- "MDC1-MRN-ATM-FANCD2 complex"
expProfile <- getCoExpProfile(complex)
}
