% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunitcoExpressionAnalysis.R
\name{getComplexmyQNXP}
\alias{getComplexmyQNXP}
\title{\code{getComplexmyQNXP} Extract expression profile for the complex subunits}
\usage{
getComplexmyQNXP(complex)
}
\arguments{
\item{complex}{(char)  name of the complex}
}
\value{
(matrix) A matrix with rows representing complex subunits in HGNC symbols
and 52 columns representing the 52 microarray experiments. Each cell of the matrix
contains the corresponded expression values
}
\description{
\code{getComplexmyQNXP} The processed expression profile from GEO is extracted
 for complex subunits in both CORUM database and GEO. For details of CORUM data
 fetch see \code{\link{fetchData}}.
}
\examples{
\dontrun{
complex <- "MDC1-MRN-ATM-FANCD2 complex"
# get the expression file of the comples subunits
myComplexQNXP <- getComplexmyQNXP(complex)
}

}
\seealso{
\code{\link{fetchData}}
}
