% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGI_PPI_Integrator.R
\name{getGeneticInteractome}
\alias{getGeneticInteractome}
\title{Return BioGrid genetic interaction tags of physical interactors of the system}
\usage{
getGeneticInteractome(mySys, criterion)
}
\arguments{
\item{mySys}{The dataframe output of \code{convertToHGNC}}

\item{criterion}{A string, either "stringent" or "relaxed", specifying whether only GGI between physical interactors should be selected.}
}
\value{
The dataframe of system components and either their GGI between physical interactors should be selected (iff \code{criterion} == "stringent")
or all GGI of physical interactors (iff \code{relaxed} == "stringent")
}
\description{
Return BioGrid genetic interaction tags of physical interactors of the system
}
