% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDirectedNetworks.R
\name{getKey}
\alias{getKey}
\title{Generates a unique key for BioGrid API access}
\usage{
getKey(my.name, my.email, my.project)
}
\arguments{
\item{my.name}{A string of the user's full name separated by space(s)}

\item{my.email}{A string matching a valid user email}

\item{my.project}{A string specifying a short project name (no spaces)}
}
\value{
The unique key for user and project for data retrieval from BioGrid
}
\description{
Generates a unique key for BioGrid API access
}
\examples{
myKey <- get_key("Tophie McGophie", "tmc\\code{at}hammertime.com", "myProj")
}
