% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunitcoExpressionAnalysis.R
\name{getpValue}
\alias{getpValue}
\title{\code{getpValue} calculate p-Value}
\usage{
getpValue(simulationCorr, myGenePairCorr)
}
\arguments{
\item{simulationCorr}{(double vector)  a vector of correlation grnerate by
randomly simulate N gene pairs and
calculate correlation.}

\item{myGenePairCorr}{(double)         gene pair coexpression correlation
that we are interested in}
}
\value{
(double) p-Value of how differentially coexpressed the gene pair is
                 over the random coexpression correlation
}
\description{
\code{getpValue} Calculate the p value of a gene pair correlation on
a simulated random gene pair coexpression corelation distribution.
}
\examples{
simulationCorr <- c(-0.5, -0.4, -0.3, -0.2, -0.1, 0, 0.1, 0.2, 0.3, 0.4, 0.5)
myGenePairCorr <- 0.1
getpValue(simulationCorr, myGenePairCorr) # 0.7630246

}
