% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyTranscriptionFactorSimilarity.R
\name{getSysTFdata}
\alias{getSysTFdata}
\title{Helper Function: get TF data only for genes in system}
\usage{
getSysTFdata(sysHGNC)
}
\arguments{
\item{sysHGNC}{list of genes from system
Accesses loaded TF data and filters for genes present in system}
}
\value{
sysTF. A smaller TF geneList with only genes in system.
}
\description{
Helper Function: get TF data only for genes in system
}
\examples{
\donttest{
getSysTFdata(sysHGNC)
}
}
