% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HuGenAnnotImport.R
\name{HuGenAnnotImport}
\alias{HuGenAnnotImport}
\title{\code{HuGenAnnotImport} imports the org.Hs.eg.db, implemented to reduce import burden of package
This database will only be imported when the semanticSimilarity.R function is required}
\usage{
HuGenAnnotImport()
}
\value{
(NULL) NULL
}
\description{
\code{HuGenAnnotImport} imports the org.Hs.eg.db, implemented to reduce import burden of package
This database will only be imported when the semanticSimilarity.R function is required
}
\examples{
\dontrun{
HuGenAnnotImport()
}
}
