% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_distance_matrices_list.R
\name{make_distance_matrices_list}
\alias{make_distance_matrices_list}
\title{Make distance matrices list}
\usage{
make_distance_matrices_list(distances, all_genes)
}
\arguments{
\item{distances}{Character vector indicating the distance metrics to be used.}

\item{all_genes}{Character vector of HGNC symbols with which to make distance matrices}
}
\value{
A list of objects of class distance.
}
\description{
\code{make_distance_matrices_list} This is a helper function for \code{\link{clusterSystems}} and is unlikely to be useful elsewhere.
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
