% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrix.R
\name{make_matrix}
\alias{make_matrix}
\title{Make Similarity Matrix}
\usage{
make_matrix(genes, sim_fn)
}
\arguments{
\item{genes}{A vector of HGNC symbols for which to calculate the similarity of all possible pairwise combinaitons.}

\item{sim_fn}{Function to be used to calculate individual pairwise similarities.}
}
\value{
similarity matrix, containing numerics values between 0 and 1
}
\description{
\code{make_matrix.R} Generate a pairwise similarity matrix for a set of genes.
}
\details{
The the similarity metric is determined by the similarity function that is passed as a parameter.
}
\section{<title>}{
 Additional explanation.
}

\examples{
# <explain what the example does>


}
\seealso{
\code{\link{<function>}} <describe related function>, ... .

Other <optional description of family>: \code{\link{expr_sim}},
  \code{\link{net_sim}}, \code{\link{tf_sim}}
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
\concept{<optional description of family>}
