% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_venn_diagram.R
\name{make_venn_diagram}
\alias{make_venn_diagram}
\title{Make Venn Diagrams}
\usage{
make_venn_diagram(system_genes, sys_name, cluster_genes, clust_num,
  jaccard, p)
}
\arguments{
\item{system_genes}{vector of genes in the system}

\item{sys_name}{name of the system}

\item{cluster_genes}{vector of genes in the cluster}

\item{clust_num}{Numeric identifier for the cluster}

\item{jaccard}{The jaccard index of the set overlap of cluster and system}

\item{p}{P value of the Jaccar index}
}
\description{
\code{make_venn_diagram} is a helper function for \code{\link{clusterSystems}} and is unlikely to be useful elsewhere.
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
