% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyTranscriptionFactorSimilarity.R
\name{makeDistanceMatrix}
\alias{makeDistanceMatrix}
\title{makeDistanceMatrix
Helper function to make binary presence absence matrix}
\usage{
makeDistanceMatrix(sysTF)
}
\arguments{
\item{sysTF}{{dataframe} output of getsysTFdata()
Makes a binary presence absence matrix which can be transformed into a
distance matrix.}
}
\value{
matrix {dataframe} A matrix with rownames as TF and column names as system genes
In the matrix 1 means TF is associated with given gene. 0 means unassociated
}
\description{
makeDistanceMatrix
Helper function to make binary presence absence matrix
}
\examples{
\donttest{
makeDistanceMatrix(list)
}
}
\author{
\href{https://orcid.org/ORCID:0000-0002-1387-487X}{Rachel Woo} (aut)
}
