% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGI_PPI_Integrator.R
\name{makeGMAP}
\alias{makeGMAP}
\title{Generate the genetic interpretation map of BioGrid GGI tags.}
\usage{
makeGMAP()
}
\value{
The dataframe mapping BioGrid GGI tag to its interpretation.
}
\description{
Generate the genetic interpretation map of BioGrid GGI tags.
}
