% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyTranscriptionFactorSimilarity.R
\name{makeMatrix}
\alias{makeMatrix}
\title{Helper function to make binary presence absence matrix}
\usage{
makeMatrix(sysTF)
}
\arguments{
\item{sysTF}{output of getsysTFdata()
Makes a binary presence absence matrix which can be transformed into a
distance matrix.}
}
\value{
matrix. A matrix with rownames as TF and column names as system genes
In the matrix 1 means TF is associated with given gene. 0 means unassociated
}
\description{
Helper function to make binary presence absence matrix
}
\examples{
\donttest{
makeMatrix(list)
}
}
