% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCorVSgotermEDA.R
\name{myPlot}
\alias{myPlot}
\title{\code{myPlot()}
Investigates relationship of expression correlation and semantic similarity of input genes
and produces plots and summary of model info}
\usage{
myPlot(geneSet)
}
\arguments{
\item{geneSet}{A set of genes of interest to investigate}
}
\value{
A heatmap of co-expression correlation of each pair of genes in the geneSet,
A heatmap of semantic similarity of each pair of genes
A scatterplot of correlation vs Go similarity
summary of linear model built by correlation vs Go similarity
}
\description{
\code{myPlot()}
Investigates relationship of expression correlation and semantic similarity of input genes
and produces plots and summary of model info
}
\examples{
geneSet <- c("AMBRA1","ATG14","ATP2A1","ATP2A2","ATP2A3")
myPlot(geneSet)

}
\author{
{Yuhan Zhang} (aut)
}
