% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_sim.R
\name{net_sim}
\alias{net_sim}
\title{Network Similarity.}
\usage{
net_sim(gene1, gene2, graph = STRINGgraph)
}
\arguments{
\item{gene1}{Character, HGNC symbol of the first gene.}

\item{gene2}{Character, HGNC sysmbol of the second gene}

\item{graph}{igraph object representing the network}
}
\value{
Numeric, similarity score: Jaccard similarity
}
\description{
\code{net_sim} Calculate the pairwise similarity of genes in an interaction network.
}
\details{
Jaccard similarity of 2 nodes in the gene network is calculated. Jaccard similarity is the number of common neighbors divided by the number of vertices that are neighbors of at least one of the 2 nodes.
Note that this functions requires the igraph package to previously be loaded.
}
\section{<title>}{
 additional explanation
}

\examples{
# Calculate the network similarity of 2 related genes:
net_sim("BRCA1", "BRCA2")
[1] 0.3481481

# Calculate network similarity of 2 genes that are not known to be related:
net_sim("ROBO1", "BRCA1")
[1] 0.01592357

}
\seealso{
\code{\link{<function>}} <describe related function>, ... .

Other <optional description of family>: \code{\link{expr_sim}},
  \code{\link{make_matrix}}, \code{\link{tf_sim}}
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
\concept{<optional description of family>}
