% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunitcoExpressionAnalysis.R
\name{pairExpHist}
\alias{pairExpHist}
\title{\code{pairExpHist} Generate distribution of coexpression correlations between
                   randomly chosen gene pairs, compared to the distribution of
                   shuffled expression values for each pair of significantly
                   differetially coexpresssed genes}
\usage{
pairExpHist(expProfile)
}
\arguments{
\item{expProfile}{(data frame)  gene pairs with co expression correlation represented
in a data frame}
}
\value{
(data frame) If there are significantly coexpressed gene pairs,
                     extract the expression profile data frame with two genes
                     that are significantly differentially coexpressed, their
                     correlation and p-value for their coexpression. Otherwise,
                     return the above information for all the gene pairs
}
\description{
\code{pairExpHist} Generate distribution of coexpression correlations between
                   randomly chosen gene pairs, compared to the distribution of
                   shuffled expression values for significantly differentially
                   coexpressed genes. In addition, this function displays the
                   distribution in a histogram for visualization and interpretation
}
\examples{
\dontrun{
complex <- "MDC1-MRN-ATM-FANCD2 complex"
complexExpHeatmap(complex)
expProfile <- getCoExpProfile(complex)
pairExpHist(expProfile)
}
}
