% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprTrajectory.R
\name{plotExprTrajectory}
\alias{plotExprTrajectory}
\title{Title.}
\usage{
plotExprTrajectory(overlayName)
}
\arguments{
\item{overlayName}{The specific Gene/Protein in the .fcs file you would like to color on the trajectories}
}
\value{

}
\description{
\code{plotExprTrajectory} Plots an iGraph of the p-Creode with coloring of the gene expression
}
\details{
Details.
}
\section{Dependency}{
 the pcreode package, which can be installed with:
\code{sudo pip install pcreode}
}

\examples{
# plots the CAR2 gene expression for the current expression trajectory
plotExprTrajectory("CAR2")

}
\author{
Yoonsik Park (au)
}
