% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelations.R
\name{postDraw}
\alias{postDraw}
\title{Helper function for the co-expression correlation & semantic similarity correlation heat maps}
\usage{
postDraw(fileName)
}
\arguments{
\item{fileName}{- if not NULL - it will call dev.off to save the file
otherwise - it prompts for user input - so the user can see the graph
before the next one is drawn}
}
\value{
(NULL)
}
\description{
Helper function for the co-expression correlation & semantic similarity correlation heat maps
}
\author{
\href{https://orcid.org/0000-0002-8778-6442}{Denitsa Vasileva} (aut)
}
