% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelations.R
\name{preDraw}
\alias{preDraw}
\title{\code{preDraw} Helper function.
Helper function for the co-expression correlation and semantic similarity correllation heat maps}
\usage{
preDraw(fileName)
}
\arguments{
\item{fileName}{(character)     if not NULL - it will use the Cairo package to create a
device. It checks the file extension to
instantiate appropriate device;
otherwise - it does nothing (graphs are plotted on screen)}
}
\value{
(NULL)
}
\description{
\code{preDraw} Helper function.
Helper function for the co-expression correlation and semantic similarity correllation heat maps
}
\author{
\href{https://orcid.org/0000-0002-8778-6442}{Denitsa Vasileva} (aut)
[Cairo package](https://CRAN.R-project.org/package=Cairo
}
