% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunitcoExpressionAnalysis.R
\name{processmyQNXP}
\alias{processmyQNXP}
\title{\code{processmyQNXP} Load and process expression-profiles}
\usage{
processmyQNXP()
}
\value{
(matrix) A matrix with 21063 rows representing genes in HGNC symbols
and 52 columns representing the 52 microarray experiments. Each cell of the matrix
contains the corresponded expression values
}
\description{
\code{processmyQNXP} The expression profiles were compiled and normalized by
using GEO data containing 52 microarray experiments.
For details see \code{\link{fetchData}}. The expression profiles were then
processed to eliminate genes with no data (row with NAs cross all experiments)
}
\examples{
\dontrun{
# fetch and process expression profile data
myQNXPProcesed <- processmyQNXP()
}

}
\seealso{
\code{\link{fetchData}}
}
