% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchData.R
\name{produceGeneNetwork}
\alias{produceGeneNetwork}
\title{\code{produceGeneNetwork} produces gene network graph for genes from the given system of genes 'sys'.}
\usage{
produceGeneNetwork(sys, gene="VPS41")
}
\arguments{
\item{sys}{(character)       5-character system code of genes}

\item{gene}{(character)      genes code}
}
\value{
a network graph
}
\description{
\code{produceGeneNetwork} outputs a graph on a network graph corresponding to a selected gene
in a system, it shows the correlation between one gene to the rest in a system.
}
\examples{
produceGeneNetwork("PHALY", "VPS41")
}

