% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchData.R
\name{produceGenesNetwork}
\alias{produceGenesNetwork}
\title{\code{produceGenesNetwork} produces genes network graph for genes from the given system of genes 'sys'.}
\usage{
produceGeneNetwork(sys, genes=c("ARF5","SPTBN2","KTF22"))
}
\arguments{
\item{sys}{(character)       5-character system code of genes}

\item{genes}{(character)      genes code}
}
\value{
a network graph
}
\description{
\code{produceGenesNetwork} outputs a graph on a network graph corresponding to a selected gene
in a system, it shows the correlation between one gene to the rest in a system.
}
\examples{
produceGenesNetwork("PHALY", c("ARF5","SPTBN2","KTF22"))
}

