% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyTranscriptionFactorSimilarity.R
\name{quantifyTFSimilarity}
\alias{quantifyTFSimilarity}
\title{quantifyTFSimilarity
Main controller function for this file. Effects all other functions.}
\usage{
quantifyTFSimilarity(filepath)
}
\arguments{
\item{filepath}{{string} to system excel file
Makes distance matrix, and cluster of shared TF presence for proteins in
the filepath denoted by system}
}
\value{
{null} writes dendrogram of TF to plots
}
\description{
quantifyTFSimilarity
Main controller function for this file. Effects all other functions.
}
\examples{
\donttest{
quantifyTFSimilarity("BCB420-2019-System-PHALY-0.3.xlsx")
}
}
\author{
\href{https://orcid.org/ORCID:0000-0002-1387-487X}{Rachel Woo} (aut)
}
