% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGI_PPI_Integrator.R
\name{recoverIDs}
\alias{recoverIDs}
\title{Try to recover IDs for ensp to sym mapping from biomart}
\usage{
recoverIDs(ensp, mart)
}
\arguments{
\item{ensp}{(character) a vector of ensemble peptide IDs}

\item{mart}{(Mart)      an ensemble mart object
"HGNC" must exist in the global namespace}
}
\value{
a dataframe with columns "ensp" containing the ensemble
peptide IDs of the input that could be mapped, and "sym",
 which contains the corresponding HGNC symbols, and rownames ensp.
}
\description{
Try to recover IDs for ensp to sym mapping from biomart
}
