% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticSimilarity.R
\name{semanticSimilarity}
\alias{semanticSimilarity}
\title{\code{semanticSimilarity} gets semantic similarity matrix of enriched
GO terms for a systems gene set}
\usage{
semanticSimilarity(sys)
}
\arguments{
\item{sys}{- System Code (e.g., PHALY)}
}
\value{
(matrix) a matrix of enriched GO terms with semantic similarity values or null if system does not contain any genes (GOterm x GOterm format for the matrix)
}
\description{
\code{semanticSimilarity} gets semantic similarity matrix of enriched
GO terms for a systems gene set
}
\examples{
\dontrun{
semanticSimilarity("PHALY")
}
}
\seealso{
Documentation of semantic similarity calculations and GO enrichment analysis \code{\link{GOSemSim}}
}
\author{
\href{https://orcid.org/0000-0003-4609-4965}{Cathy Cha} (aut)
}
