% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqComparisonTable.R
\name{SeqComparisonTable}
\alias{SeqComparisonTable}
\title{the gene it is being compared to, the protein sequence of
    each resulting protein generated by the genes and
    the consensus sequence and conservation scores based
    off a ClustalW multiple Sequence Alignment.}
\usage{
SeqComparisonTable(hgnc)
}
\arguments{
\item{hgnc}{The hgnc symbol of the gene of interest in the form of a
string. This hgnc gene will have multiple sequence
alignments being performed on it with all of it's interacting
proteins based off the STRING database.}
}
\value{
A data frame including the hgnc gene of interest,
    the gene it is being compared to, the protein sequence of
    each resulting protein generated by the genes and
    the consensus sequence and conservation scores based
    off a ClustalW multiple Sequence Alignment.
}
\description{
\code{SeqComparisonTable} Return a data frame containing the alignment data
   between the provided hgnc gene and it's corresponding genes based
   on the STRING database..
}
\examples{
# Retrieve system database using functions provided from the master package
myDB <- fetchData("SysDB")
# See which genes the "PHALY" systems contains
(SyDBgetSysSymbols(myDB, "PHALY"))
# Pick a sample HGNC symbol within the PHALY system
hgnc <- "BECN1"
# Call the SeqComparisonTable function to generate the dataframe
seqComparisonDF <- SeqComparisonTable(hgnc)

}
\author{
{Heewon Lee} (aut)
}
