% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceAnalysisESA.R
\name{sequenceAnalysis}
\alias{sequenceAnalysis}
\title{\code{sequenceAnalysisESA} Return a data frame containing the alignment data
   between the provided hgnc gene and it's corresponding genes based
   on the STRING database..}
\usage{
sequenceAnalysis(hgnc)
}
\arguments{
\item{hgnc}{The hgnc symbol of the gene of interest in the form of a
string. This hgnc gene will have multiple sequence
alignments being performed on it with all of it's interacting
proteins based off the STRING database.}
}
\value{
A data frame including the hgnc gene of interest,
    the gene it is being compared to, the protein sequence of
    each resulting protein generated by the genes and
    the consensus sequence and conservation scores based
    off a ClustalW multiple Sequence Alignment.
}
\description{
\code{sequenceAnalysisESA} Return a data frame containing the alignment data
   between the provided hgnc gene and it's corresponding genes based
   on the STRING database..
}
\examples{
# Picking a sample hgnc symbol to get the corresponding protein alignments
# with it's interacting genes
hgnc <- "BECN1"

# Call the getAlignments helper function to generate the dataframe
result <- getAlignments(hgnc)
# If the user would like, the resulting data frame can be stored
# in a .tsv file which can be imported into the environment by the
# user

}
\author{
{Heewon Lee} (aut)
}
