% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRINGExplore.R
\name{STRINGExplore}
\alias{STRINGExplore}
\title{STRINGExplore}
\usage{
STRINGExplore(geneSys)
}
\arguments{
\item{geneSys}{(str) A 5 character string representing the name of
the desired curated gene system user would like to build a
high-confidence interaction network for.}
}
\value{
A labeled list("nodes", "edges", "network") containing the visNetwork
object and the edge and node dataframes resulting from the setup of the network.
May be useful for downstream analysis by the end user.
}
\description{
\code{STRINGExplore} Utilizes the combined score and protein action information
 available on STRING to create an interactive network of high confidence (>0.8) interaction
}
\section{Details}{
 The STRING edges on their don't provide a good starting
 point for hypothesizing about the reasons for the existence of high-
 confidence interactions. By combining the easy-to-navigate, interactive
 network from the visNetwork package with tooltips of information from
 high-confidence (>0.8) data from STRING protein_actions dataset, STRINGExplore
 aims to ease the process of deciding on future research avenues. The graph layout takes
 into consideration the betweeness centrality of nodes, and so when inputting a gene list
 from a curated system it's highly likely that highly-centrality nodes will impart some
 important functionality to the system.
}

\examples{
STRINGExplore("PHALY")

}
\author{
\href{https://orcid.org/0000-0003-4762-8797}{Gabriela Morgenshtern} (aut)
}
