% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyDButils.R
\name{SyDBgetIDforKey}
\alias{SyDBgetIDforKey}
\title{\code{SyDBgetIDforKey} fetch an ID from a systems database.}
\usage{
SyDBgetIDforKey(val, att, tbl, db, v = "2.1.1")
}
\arguments{
\item{val}{(character)  match value "val" ...}

\item{att}{(character)  ... in attribute column "att" ...}

\item{tbl}{(character)  ... of table "tbl" ...}

\item{db}{(character)  in database "db".}

\item{v}{(character)   database version. Defaults to 2.1.1}
}
\value{
(character) a vector of HGNC symbols or a zero-length vector if
                    none were found
}
\description{
\code{SyDBgetIDforKey} documentation in progress ...
}
\examples{
\dontrun{
SyDBgetIDforKey("PHALY","code","system", mySDB) # get the ID for the system
                                               # with code "PHALY"
}
}
