% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyDButils.R
\name{SyDBgetSysSymbols}
\alias{SyDBgetSysSymbols}
\title{\code{SyDBgetSysSymbols} fetch root-system IDs from a systems database.}
\usage{
SyDBgetSysSymbols(db, sysCode, treatNOT = "exclude", v = "2.1.1")
}
\arguments{
\item{db}{(character)  name of the database}

\item{sysCode}{(character)  the requested system code}

\item{treatNOT}{(character)  not implemented}

\item{v}{(character)  the database version. Defaults to 2.1.1}
}
\value{
(character) a vector of HGNC symbols or a zero-length vector if
                    none were found
}
\description{
\code{SyDBgetSysSymbols} documentation in progress ...
}
\examples{
\dontrun{
SyDBgetSysSymbols("PHALY")            # all HGNC symbols in the PHALY system
}
}
