% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyDButils.R
\name{SyDBgetValForID}
\alias{SyDBgetValForID}
\title{\code{SyDBgetValForID} fetch an an attribute value for an ID from a systems database.}
\usage{
SyDBgetValForID(id, att, tbl, db, all = FALSE)
}
\arguments{
\item{id}{(character)  return values matching IDs in "id" ...}

\item{att}{(character)  ... from column "att" ...}

\item{tbl}{(character)  ... of the table "tbl" ...}

\item{db}{(character)  in database "db".}

\item{all}{(logical)  if TRUE, then NA will be added to the returned
vector, one for each element in "id" that was
not found. If FALSE (default), only values for
matching IDs will be returned, i.e. the length
of the input is not preserved.}
}
\value{
(character) a vector of values, or NA where
                    none were found
}
\description{
\code{SyDBgetValFforID} documentation in progress ...
}
\examples{
\dontrun{
# get the IDs for the systems with code "notAnID" and "PHALY". Element 1 is NA.
myIDs <- c("notAnID", "give.jams-1d8-648b-1e12-6a9f-65421424affe")
SyDBgetValForID(myIDs, "code", "system", mySDB, all = TRUE)
}
}
