% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyDButils.R
\name{SyDBinvalid}
\alias{SyDBinvalid}
\title{\code{SyDBinvalid} validate a systems database.}
\usage{
SyDBinvalid(db)
}
\arguments{
\item{db}{(character)  s systems database "db".}
}
\value{
(character) an error message describing the problem if the database
                    is invalid, or an empty string if the database is valid.
}
\description{
\code{SyDBinvalid} documentation in progress. Tests correct type, structure,
                   and version of the database ...
}
\examples{
\dontrun{
# use the following idiom in functions:
if (as.logical(nchar(msg <- SyDBinvalid(<database parameter>)))) {stop(msg)}
}
}
