% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symComp.R
\name{symComp}
\alias{symComp}
\title{\code{symComp} Return a list of domains that the genes in two systems
   overlap along with genes that mapped to the same domains}
\usage{
symComp(system1, system2)
}
\arguments{
\item{system1}{The name of the system in interest. Is compared with system2
in this function.}

\item{system2}{The name of the system in interest. Is compared with system1
in this function.}
}
\value{
A list of overlapping Pfam domains along with genes that mapped to them.
    Also, a horizonal bar graph is generated to summarize the count of genes
    mapped to each overlapping domains.
}
\description{
\code{symComp} Return a list of domains that the genes in two systems
   overlap along with genes that mapped to the same domains
}
\examples{
# Picking sample system1 and system2 to get intersect domains of the
two systems
# Call the symComp helper function to generate list
result <- symComp("ANPRC", "ANPRC2")

}
\author{
{Tina Lee} (aut)
}
