% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDepletTF.R
\name{sysCorGenes}
\alias{sysCorGenes}
\title{sysUpCorGenes.}
\usage{
sysCorGenes(exProfS, nUpMost = FALSE, direction = "Positive",
  multipleTests = "Bonferroni", alpha = 0.05)
}
\arguments{
\item{exProfS}{(numeric)    A matrix of expression profiles, with column dimensions
equal to the number of microarray experiments, and row
dimensions that are equal to the number of queried genes.
All genes must have information for all experiments (i.e.,
no NAs are allowed). The matrix can be produced  by the
helper function sysExProf().}

\item{nUpMost}{(numeric|logical) Either FALSE or a numeric vector of 1L length. FALSE
returns all significantly correlated genes. A numeric
vector sets the number of most highly correlated
pairs of genes to return by the function. The default
is FALSE.}

\item{direction}{(character) A vector of 1L length of either "Positive" or "Negative".
"Positive" returns pairs of genes that are significantly
positively correlated. "Negative" returns pairs of genes
that are significantly negatively correlated. The
default is set to "Positive".}

\item{multipleTests}{(character) A vector of 1L length of either "Bonferroni" or
"BH" (Benjamini-Hochberg). Sets the multiple tests
correction approach. The default is "Bonferroni".}

\item{alpha}{(numeric)      A vector of 1L length. Sets the value of alpha for
deciding the cutoff for significant results. The Default
is 0.05.}
}
\value{
(character)         A vector of unique genes (HGNC symbols) that were found to
                            be significantly correlated, according to the conditions
                            set by the parameters. For a numeric value of nUpMost,
                            although the number of pairs of genes was set, the number
                            of the returned genes is not always equal to 2 * "nUpMost",
                            because some of the pairs may contain the same genes,
                            while the function returns only a vector of unique genes.
}
\description{
\code{sysUpCorGenes} Finding pairs of genes that their expression profiles are the
                     most positively or negatively correlated.
}
\section{Details}{
 A helper function that produces a character vector of unique genes
                  (HGNC symbols) that belongs to pairs of genes that their expression
                  profiles were shown to be either positively or negatively correlated.
                  It is also possible to specify the number the of most highly
                  negatively or positively correlated pairs of genes to retrieve.
}

\examples{
\dontrun{
# Vector of the 12 up most positively correlated pairs of genes of "SLIGR"
# system.
# mySymbols <- SyDBgetSysSymbols(fetchData("SysDB"),"SLIGR")[[1]]
# exProfS <- matrix(data = sample(1:200, 4 * length(mySymbols), replace = TRUE),
#                  ncol = 4) #generating synthetic expression data
# row.names(exProfS) <- mySymbols
# sysUpCorGenes(exProfS = exProfS, nUpMost = 12)
}
}
\seealso{
\code{\link{cor.test}} Test for association between paired samples, using one
                                of Pearson's product moment correlation coefficient,
                                Kendall's tau or Spearman's rho.
}
\author{
\href{https://orcid.org/0000-0002-9478-5974}{Sapir Labes} (aut)
}
