% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDepletTF.R
\name{sysExProf}
\alias{sysExProf}
\title{sysExProf.}
\usage{
sysExProf(GeneSym)
}
\arguments{
\item{GeneSym}{(character)   A vector of length > 0L of HGNC symbols.}
}
\value{
(numeric)           A matrix with row dimensions that are equal to the
                             number of genes of the GeneSym vector that has
                             expression data for all of the experiments.
                             The column dimensions corresponds to the number of
                             microarray experiments in the expression profile
                             dataset. The matrix contains the expression profiles
                             of genes that have data for all experiments in the
                             expression profile source dataset. The rows names
                             are the HGNC sybmols and the column names are the
                             experiments' names.
}
\description{
\code{sysExProf} Loads expression-profiles of chosen genes
}
\section{Details}{
 This is a helper function. The expression profiles were compiled
from microarray experiments downloaded from GEO, and quantile normalized. For the
function that fetches the database, see \code{\link{fetchData}}.
}

\examples{
\dontrun{
# Expression profiles for "BRCA1" and "AR" genes.
# sysExProf(GeneSym = c("BRCA1", "AR"))
}
}
\seealso{
\code{\link{fetchData}} Fetches the requested dataset.
}
\author{
\href{https://orcid.org/0000-0002-9478-5974}{Sapir Labes} (aut)
}
