% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDepletTF.R
\name{sysGTRDgenes}
\alias{sysGTRDgenes}
\title{sysGTRDgenes.}
\usage{
sysGTRDgenes(GeneSym)
}
\arguments{
\item{GeneSym}{(character)  A vector of length > 0L of HGNC symbols.}
}
\value{
(list)              A list of HGNC-symbol-named character vectors of
                            the transcription factors that bind each gene.
}
\description{
\code{sysGTRDgenes} Fetch GTRD data for the specified genes.
}
\section{Details}{
 This is a helper function. The fetched dataset was composed from the
GTRDgeneTFs dataset which lists which TFs bind the upstream regulatory regions of each
specified gene, as observed by ChIP-seq experiments. For more details about the
function that fetches the GRTDgeneTFs database, see \code{\link{fetchData}}.
}

\examples{
# Fetching the vector of the transcription factors that bind the gene BCL3.
# sysGTRDgenes(GeneSym = c("BCL3"))

}
\seealso{
\code{\link{fetchData}} Fetches the requested dataset.
}
\author{
\href{https://orcid.org/0000-0002-9478-5974}{Sapir Labes} (aut)
}
