% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnrichDepletTF.R
\name{sysGTRDtf}
\alias{sysGTRDtf}
\title{sysGTRDtf.}
\usage{
sysGTRDtf(TfSym, GeneSym)
}
\arguments{
\item{TfSym}{(character)     A vector of length > 0L of unique TFs (HGNC symbols)
that was extracted from the output of sysGTRDgenes()
for a certain set of genes.}

\item{GeneSym}{(character)   A vector of length > 0L of the unique genes (HGNC symbols)
that the function sysGTRDgenes() retrieved for an input
of a certain set of genes (the same certain set of genes
that was used to produce TfSym).}
}
\value{
(list)               A list of transcription-factor-named character vectors of
                             the genes that bind each transcription factor.
}
\description{
\code{sysGTRDtf} Fetch GTRD data for the specified Transcription Factors and genes.
}
\section{Details}{
 This is a helper function. The fetched dataset was collected and
composed from the GTRD database. The GTRDTFgenes dataset lists which  genes bind
each specified TF, as observed by ChIP-seq experiments. For more details about the
function that fetches the GTRDTFgenes database, see \code{\link{fetchData}}. This
function requires an input that can be produced by the helper function sysGTRDgenes.
}

\examples{
\dontrun{
# Fetching the vector of the genes that bind the transcription factor AR.
GTRDgenes <- sysGTRDgenes(GeneSym = c("LAMP1", "HDAC6"))
mySymbols <- names(GTRDgenes) #The specific set of genes retrieved
myTfSym <- unique(unlist(GTRDgenes, use.names = FALSE)) #TFs that bind the specific set of genes
sysGTRDtf(TfSym = myTfSym, GeneSym = mySymbols)
}
}
\seealso{
\code{\link{fetchData}} Fetches the requested dataset.
}
\author{
\href{https://orcid.org/0000-0002-9478-5974}{Sapir Labes} (aut)
}
