% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/System_Expression.R
\name{System_Expression}
\alias{System_Expression}
\title{System level expression values.}
\usage{
System_Expression(sysname, myQNXP)
}
\arguments{
\item{sysname}{(character) 5-letter code of system being investigated}

\item{myQNXP}{(matrix) Matrix of expressions. Columns represent experiments and
rows individual genes}
}
\value{
(list) A list containing 3 elements. The first the mean log normalized
gene expression by condition. Second and third the conditions where the system
is significantly up and down regulated respectively.
}
\description{
\code{exProf} Returns the log-normalized expression values for gene provided.
}
\details{
\code{System_Expression} Provides the mean expression level for each experiment
present in the dataset. As well, displays which conditions are up and down
regulted significantly.

For further information, see details in the \href{../doc/BCB420.2019.ESA-Vignette_McNeil.html}{ System Expression vignette} (or load the vignette with \code{vignette("BCB420.2019.ESA-Vignette_McNeil", package = "BCB420.2019.ESA")}).
}
\author{
\href{https://orcid.org/0000-0002-2821-6204}{Matthew McNei;} (aut)
}
