% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESA.R
\name{systemAnalysis}
\alias{systemAnalysis}
\title{\code{systemAnalysis} Takes in a biological system and a gene from that system
picked by the user and and produces a histogram for the node degree distribution
of the system and a network plot for the gene from the system.}
\usage{
systemAnalysis(gene, sys)
}
\arguments{
\item{gene}{A gene from a Biological system}

\item{sys}{A Biological system}
}
\description{
\code{systemAnalysis} Takes in a biological system and a gene from that system
picked by the user and and produces a histogram for the node degree distribution
of the system and a network plot for the gene from the system.
}
\examples{
\dontrun{
# Retrieve a histogram for the node degrees of the PHALY gene network and
# get the network plot of the AMBRA1 gene from the PHALY system
systemAnalysis("AMBRA1", "PHALY")
}

}
\author{
{Deus Bajaj} (aut)
}
