% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_dist.R
\name{tf_dist}
\alias{tf_dist}
\title{Transcription Factor Distance}
\usage{
tf_dist(gene1, gene2, geneList)
}
\arguments{
\item{gene1}{String, HGNC symbol for the first gene.}

\item{gene2}{String, HGNC symbol for the second gene.}

\item{geneList}{A list of lists of transcription factors than have binding sites upstream of genes. Generated by calling \code{\link{fetchData}("GTRDgeneTFs")}.}
}
\value{
The similarity score of gene1 and gene2: number of shared upstream transcription factors.
}
\description{
\code{tf_dist} Calculate upstream transcription factor binding site distace between 2 genes.
}
\details{
Calculate the distance/dissimilarity of upstream transcription factors of 2 genes.
Similarity is taken to be the jaccard index of the set overlap between the 2 transcription factors sets. Distance is then taken to be 1 - similarity.
}
\examples{
# Calculate the transcriptio factor distance of 2 related genes "BRCA1" and "BRCA2"
geneList <- fetchData("GTRDgeneTFs")
tf_dist("BRCA1", "BRCA2", geneList)


}
\seealso{
\code{\link{fetchData}} For format of geneList
}
\author{
\href{https://orcid.org/0000-0001-5724-2252}{Rachel Silverstein} (aut)
}
