% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoSymComp.R
\name{twoSymComp}
\alias{twoSymComp}
\title{\code{twoSymComp} Return a list of domains that the genes in two systems
   overlap along with genes that mapped to the same domains}
\usage{
twoSymComp(system1, system2)
}
\arguments{
\item{system1}{(character) The name of the system in interest. Is compared with system2
in this function.}

\item{system2}{(character) The name of the system in interest. Is compared with system1
in this function.}
}
\value{
(list) A list of overlapping Pfam domains along with genes that mapped to them.
    Also, a horizonal bar graph is generated to summarize the count of genes
    mapped to each overlapping domains.
}
\description{
\code{twoSymComp} Return a list of domains that the genes in two systems
   overlap along with genes that mapped to the same domains
}
\examples{
# Picking sample system1 and system2 to get intersect domains of the
# two systems
# Call the twoSymComp helper function to generate list
result <- twoSymComp("PHALY", "SLIGR")

}
\author{
{Tina Lee} (aut)
}
