% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyTranscriptionFactorSimilarity.R
\name{visualize}
\alias{visualize}
\title{Helper Visualization Function}
\usage{
visualize(matrix)
}
\arguments{
\item{matrix}{in the form of output of makeMatrix()
Function which makes distanceMatrix, cluster and exports dendrogram}
}
\value{
distance binary matrix and dendrogram in plots
}
\description{
Helper Visualization Function
}
\examples{
\donttest{
visualize(matrix)
}
}
