% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyTranscriptionFactorSimilarity.R
\name{visualizeDendrogram}
\alias{visualizeDendrogram}
\title{visualizeDendrogram
Helper Visualization Function}
\usage{
visualizeDendrogram(matrix)
}
\arguments{
\item{matrix}{{dataframe} in the form of output of makeDistanceMatrix(). Binary matrix
Function which makes distanceMatrix, cluster and exports dendrogram}
}
\value{
{null} writes dendrogram in plots
}
\description{
visualizeDendrogram
Helper Visualization Function
}
\examples{
\donttest{
visualizeDendrogram(matrix)
}
}
\author{
\href{https://orcid.org/ORCID:0000-0002-1387-487X}{Rachel Woo} (aut)
}
