% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGI_PPI_Integrator.R
\name{visualizeInteractions}
\alias{visualizeInteractions}
\title{A helper function of \code{hypothesize} that plots the hypothesis graph}
\usage{
visualizeInteractions(network, emap, ppi_ggi, gmap)
}
\arguments{
\item{network}{A dataframe of physically-interacting system components with their genetic interactions and}

\item{emap}{The dataframe obtained from \code{makeEMAP}.}

\item{ppi_ggi}{An optional dataframe to specify subset of \code{mySys} for which both PPI and GGI data is available}

\item{gmap}{The dataframe obtained from \code{makeGMAP}.}
}
\description{
A helper function of \code{hypothesize} that plots the hypothesis graph
}
