% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDirectedNetworks.R
\name{convertToHGNC}
\alias{convertToHGNC}
\title{Convert a system's components to HGNC-annotated physical interactors}
\usage{
convertToHGNC(myGenes, mart)
}
\arguments{
\item{myGenes}{A list of system's ENSEMBL peptide IDs}

\item{mart}{A biomaRt mart to be queried for ENSEMBL ID conversion}
}
\value{
The dataframe of physically interacting genes in \code{myGenes} according to \code{mart}
}
\description{
Convert a system's components to HGNC-annotated physical interactors
}
\examples{
ensembl <- useMart(biomart = "ensembl")
human <- searchDatasets(mart = ensembl, pattern = "hsapiens")
myMart <- useMart("ensembl", human$dataset)
PHALY <- c("AMBRA1", "ATG14", "ATP2A1", "ATP2A2", "ATP2A3", "BECN1", "BECN2",
"BIRC6", "BLOC1S1", "BLOC1S2", "BORCS5", "BORCS6", "BORCS7",
"BORCS8", "CACNA1A", "CALCOCO2", "CTTN", "DCTN1", "EPG5", "GABARAP",
"GABARAPL1", "GABARAPL2", "HDAC6", "HSPB8", "INPP5E", "IRGM")
mySys <- convertToHGNC(PHALY, myMart)
}
