% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainEnrichment.R
\name{domainEnrichment}
\alias{domainEnrichment}
\title{Domain enrichment in a system.}
\usage{
domainEnrichment(system = "PHALY")
}
\arguments{
\item{system}{The shortehand notation of selected system, five-letter code as discussed in class.}
}
\value{
<description>.
}
\description{
\code{domainEnrichment} Find the enriched domains in selected system and interpret enrichment by plot.
}
\details{
Details.
}
\section{<title>}{
 Additional explanation.
}

\examples{
system <- "PHALY"
domainEnrichment(system)

}
\author{
\href{https://orcid.org/0000-0001-8720-5874}{Fan Shen} (aut)
}
