% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchData.R
\name{fetchData}
\alias{fetchData}
\title{\code{fetchData} read a prepared dataset from a repository.}
\usage{
fetchData(set)
}
\arguments{
\item{set}{(character)  name of the requested dataset}
}
\value{
a dataset
}
\description{
\code{fetchData} documentation forthcoming.
}
\examples{
fetchData()                         # prints avaialable datasets
HGNC <- fetchData("HGNCreference")  # assigns the HGNC reference dataset
}
