% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSTRINGKnodes.R
\name{findSTRINGKnodes}
\alias{findSTRINGKnodes}
\title{\code{findSTRINGKnodes} output the K node values for vertices in the STRING interaction graph given the system of genes 'sys'.}
\usage{
findSTRINGKnodes(sys)
}
\arguments{
\item{sys}{(character)       5-character system code of genes}
}
\value{
(numeric) Knode values
}
\description{
\code{findSTRINGKnodes} outputs a the K node values (as defined by Cornish & Markowetz, 2014),
a measure of the vertex association to the high-weighted vertices defined by the input
curated gene system 'sys'. Also plots the knode the top \emph{n} knode values
}
\examples{
# Find the Knodes values for genes in the STRING graph sorted by association to PHALY genes
findSTRINGKnodes("PHALY")
}
