% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneticInteractome.R
\name{getGeneticInteractome}
\alias{getGeneticInteractome}
\title{getGeneticInteractome}
\usage{
getGeneticInteractome(mySys, criterion)
}
\arguments{
\item{mySys}{The dataframe of PPI between system's components}

\item{criterion}{A string, either "stringent" or "relaxed", specifying
whether only GGI between physical interactors should be selected.}
}
\value{
The dataframe of system components and either their GGI between
physical interactors should be selected (iff \code{criterion} == "stringent")
or all GGI of physical interactors (iff \code{relaxed} == "stringent")
}
\description{
getGeneticInteractome
}
\author{
Nada Elnour, \email{nada.elnour@mail.utoronto.ca}

\code{getGeneticInteractome} augments systems with BioGRID genetic
interaction annotations
}
