% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneticInteractome.R
\name{getGeneticInteractome}
\alias{getGeneticInteractome}
\title{getGeneticInteractome}
\usage{
getGeneticInteractome(mySys, criterion)
}
\arguments{
\item{mySys}{(dataframe) A 2-column dataframe of PPI between system's
components.}

\item{criterion}{(string) Either "stringent" or "relaxed", specifying
whether only GGI between physical interactors should be selected.}
}
\value{
(dataframe) A 3-column dataframe of PPIs and either
the GGIs between them (iff \code{criterion} == "stringent") or
all GGIs that implicate them (iff \code{relaxed} == "stringent"). The first
two columns denote the interacting pair; the third is the type of genetic
interaction.
}
\description{
getGeneticInteractome
}
\examples{
\dontrun{
library(data.table)
biogrid <- fetchData("BioGRID")
url <- paste0("http://www.informatics.jax.org/go/report.txt?",
              "goID=GO:0006099&results=42&startIndex=0&sort=term&dir=")
myGenes <- fread(url)
myGenes <- myGenes$`MGI Gene/Marker ID`
myGenes <- toupper(myGenes)
PPI <- biogrid[biogrid$type == "physical", ]
sel <- PPI$A \%in\% myGenes & PPI$B \%in\% myGenes
mySys <- PPI[sel, ]
ppi_ggi <- getGeneticInteractome(mySys, "stringent")
network <- getGeneticInteractome(mySys, "relaxed")
hypothesize(network, ppi_ggi)
}
}
\author{
Nada Elnour, \email{nada.elnour@mail.utoronto.ca}

\code{getGeneticInteractome} augments systems with BioGRID genetic
interaction annotations
}
