% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDirectedNetworks.R
\name{getGeneticInteractome}
\alias{getGeneticInteractome}
\title{Return BioGrid genetic interaction tags of physical interactors of the system}
\usage{
getGeneticInteractome(mySys, criterion)
}
\arguments{
\item{mySys}{The dataframe output of \code{convertToHGNC}}

\item{criterion}{A string, either "stringent" or "relaxed", specifying whether only GGI between physical interactors should be selected.}
}
\value{
The dataframe of system components and either their GGI between physical interactors should be selected (iff \code{criterion} == "stringent") or all GGI of physical interactors (iff \code{relaxed} == "stringent")
}
\description{
Return BioGrid genetic interaction tags of physical interactors of the system
}
\examples{
ensembl <- useMart(biomart = "ensembl")
human <- searchDatasets(mart = ensembl, pattern = "hsapiens")
myMart <- useMart("ensembl", human$dataset)
PHALY <- c("AMBRA1", "ATG14", "ATP2A1", "ATP2A2", "ATP2A3", "BECN1", "BECN2",
"BIRC6", "BLOC1S1", "BLOC1S2", "BORCS5", "BORCS6", "BORCS7",
"BORCS8", "CACNA1A", "CALCOCO2", "CTTN", "DCTN1", "EPG5", "GABARAP",
"GABARAPL1", "GABARAPL2", "HDAC6", "HSPB8", "INPP5E", "IRGM")
mySys <- convertToHGNC(PHALY, myMart) \%>\% getGeneticInteractome(criterion = "stringent")
}
