% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyTranscriptionFactorSimilarity.R
\name{getSysGenes}
\alias{getSysGenes}
\title{Helper Function: Read system data from excel}
\usage{
getSysGenes(filepath)
}
\arguments{
\item{filepath}{to system excel sheet
Helper function to read in system data from excel}
}
\value{
sysHGNC. A list of genes in the system
}
\description{
Helper Function: Read system data from excel
}
\examples{
\donttest{
getSysGenes("BCB420-2019-System-PHALY-0.3.xlsx")
}
}
\author{
\href{https://orcid.org/ORCID:0000-0002-1387-487X}{Rachel Woo} (aut)
}
