% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGI_PPI_Integrator.R
\name{getSysInteractions}
\alias{getSysInteractions}
\title{Filter system by physically-interacting components and return their genetic interactions}
\usage{
getSysInteractions(filename, mart = myMart, criterion = "stringent")
}
\arguments{
\item{filename}{A string specifying the full path to the excel sheet containing the system's components}

\item{mart}{A biomaRt mart to be queried for ENSEMBL ID conversion}

\item{criterion}{A string, either "stringent" or "relaxed", specifying whether only GGI between physical interactors should be selected.}
}
\value{
A dataframe of system components and either their GGI between physical interactors should be selected (iff \code{criterion} == "stringent")
or all GGI of physical interactors (iff \code{relaxed} == "stringent")
}
\description{
Filter system by physically-interacting components and return their genetic interactions
}
