% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictDirectedNetworks.R
\name{getSysInteractions}
\alias{getSysInteractions}
\title{getSysInteractions}
\usage{
getSysInteractions(sysName, criterion = "stringent")
}
\arguments{
\item{sysName}{(string|vector of strings) Specifies the systems of interest}

\item{criterion}{(string) Either "stringent" or "relaxed", specifying if
only GGI between physical interactors should be selected.}
}
\value{
(dataframe) A 3-column dataframe of system components and either
their GGI between physical interactors should be selected
(iff \code{criterion} == "stringent") or all GGI of physical interactors
(iff \code{relaxed} == "stringent"). The first two columns denote the
interacting pair; the third is the type of genetic interaction.
}
\description{
getSysInteractions
}
\examples{
\dontrun{
# Get SLIGR physical interactors and return the genetic interactions between
# them
mySys <- getSysInteractions("SLIGR", criterion = "stringent")
}
}
\author{
Nada Elnour, \email{nada.elnour@mail.utoronto.ca}

\code{getSysInteractions} filter system by physically-interacting components
and return their genetic interactions
}
