% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyTranscriptionFactorSimilarity.R
\name{getSysTFdata}
\alias{getSysTFdata}
\title{getSysTFdata
Helper Function: get TF data only for genes in system}
\usage{
getSysTFdata(sysHGNC)
}
\arguments{
\item{sysHGNC}{{list of string} list of genes from system
Accesses loaded TF data and filters for genes present in system}
}
\value{
sysTF {list} A smaller TF geneList with only genes in system.
}
\description{
getSysTFdata
Helper Function: get TF data only for genes in system
}
\examples{
\donttest{
getSysTFdata(sysHGNC)
}
}
\author{
\href{https://orcid.org/ORCID:0000-0002-1387-487X}{Rachel Woo} (aut)
}
