% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{hypothesize}
\alias{hypothesize}
\title{Generates an annotated hypothesis graph whose nodes are components of \code{mySys} and edges are GGI interpretations in either \code{EMAP} or \code{GMAP}}
\usage{
hypothesize(network, ppi_ggi = NULL)
}
\arguments{
\item{network}{A dataframe of physically-interacting system components with their genetic interactions and}

\item{ppi_ggi}{An optional dataframe to specify subset of \code{mySys} for which both PPI and GGI data is available}
}
\description{
Generates an annotated hypothesis graph whose nodes are components of \code{mySys} and edges are GGI interpretations in either \code{EMAP} or \code{GMAP}
}
\examples{
mySys <- getSysInteractions("SLIGR", criterion = "stringent")
mySys2 <- getSysInteractions("SLIGR", criterion = "relaxed")
hypothesize(mySys)
hypothesize(mySys2, mySys)

}
