% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installgenome.R
\name{installHumanGenomeAnnotation}
\alias{installHumanGenomeAnnotation}
\title{\code{installHumanGenomeAnnotation} install Human genome Annotation when necessary.}
\usage{
installHumanGenomeAnnotation()
}
\value{
(NULL)
}
\description{
\code{installHumanGenomeAnnotation} install Human genome Annotation when necessary.
}
\examples{
\dontrun{
installHumanGenomeAnnotation()
}
}
