% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMaps.R
\name{makeEMAP}
\alias{makeEMAP}
\title{Generate the genetic interpretation map of BioGrid GGI tags of system's physical interactions.}
\usage{
makeEMAP()
}
\value{
The dataframe mapping BioGrid GGI tag to its interpretation assuming that the system's components also interact physically.
}
\description{
Generate the genetic interpretation map of BioGrid GGI tags of system's physical interactions.
}
\examples{
EMAP <- makeEMAP()

}
