% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMaps.R
\name{makeGMAP}
\alias{makeGMAP}
\title{Generate the genetic interpretation map of BioGrid GGI tags.}
\usage{
makeGMAP()
}
\value{
(datafram) An 11-by-2 dataframe mapping BioGrid GGI tag to its
interpretation. The first
 column contains the official BioGRID GGI tags; the second contains
 interpreted relationships.
}
\description{
Generate the genetic interpretation map of BioGrid GGI tags.
}
\examples{
GMAP <- makeGMAP() # generates and loads teh GMAP

}
