% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneCorVSgotermEDA.R
\name{myScatterPlot}
\alias{myScatterPlot}
\title{\code{myScatterPlot()} Expression and Semantic Similarity}
\usage{
myScatterPlot(geneSet)
}
\arguments{
\item{geneSet}{A set of genes of interest to investigate}
}
\value{
(list) list of graphs containing A heatmap of co-expression correlation of each pair of genes in the geneSet,
                 A heatmap of semantic similarity of each pair of genes,and
                 A scatterplot of correlation vs Go similarity with summary of linear model built by correlation vs Go similarity
}
\description{
\code{myScatterPlot()}  Investigates relationship of expression correlation and semantic similarity of input genes
and produces plots and summary of model info.
[GOSemSim package](http://bioconductor.org/packages/release/bioc/html/GOSemSim.html)
[org.Hs.eg.db data package](https://bioconductor.org/packages/release/data/annotation/html/org.Hs.eg.db.html)
}
\examples{
\dontrun{
geneSet <- c("AMBRA1","ATG14","ATP2A1","ATP2A2","ATP2A3")
myScatterPlot(geneSet)
}
}
\author{
{Yuhan Zhang} (aut)
}
