% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subunitcoExpressionAnalysis.R
\name{processmyQNXP}
\alias{processmyQNXP}
\title{\code{processmyQNXP} Load and process expression-profiles}
\usage{
processmyQNXP()
}
\value{
(matrix) A matrix with 21063 rows representing genes in HGNC symbols
and 52 columns representing the 52 microarray experiments. Each cell of the matrix
contains the corresponded expression values
}
\description{
\code{processmyQNXP} Load and process expression-profiles
}
\examples{
\dontrun{
# fetch and process expression profile data
myQNXPProcesed <- processmyQNXP()
}

}
