% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantifyTranscriptionFactorSimilarity.R
\name{quantifyTFSimilarity}
\alias{quantifyTFSimilarity}
\title{Main controller function for this file. Effects all other functions.}
\usage{
quantifyTFSimilarity(filepath)
}
\arguments{
\item{filepath}{to system excel file
Makes distance matrix, and cluster of shared TF presence for proteins in
the filepath denoted by system}
}
\value{
dendrogram of TF
}
\description{
Main controller function for this file. Effects all other functions.
}
\examples{
\donttest{
quantifyTFSimilarity("BCB420-2019-System-PHALY-0.3.xlsx")
}
}
