% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semanticSimilarity.R
\name{semanticSimilarity}
\alias{semanticSimilarity}
\title{\code{semanticSimilarity} gets semantic similarity matrix of enriched
GO terms for a systems gene set}
\usage{
semanticSimilarity(sys)
}
\arguments{
\item{sys}{- System Code (e.g., PHALY)}
}
\value{
(matrix) a matrix of enriched GO terms with semantic similarity values
}
\description{
\code{semanticSimilarity} gets semantic similarity matrix of enriched
GO terms for a systems gene set
}
\examples{
\dontrun{
semanticSimilarity("PHALY")
}
}
\author{
\href{https://orcid.org/0000-0002-1134-6758}{Boris Steipe} (aut)
}
