% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqSimVis.R
\name{seqSimVis}
\alias{seqSimVis}
\title{\code{seqSimVis()} Calculates sequence similarity and plots dot plots based on the
three highest similarity values to visualize}
\usage{
seqSimVis(gene_of_int, system)
}
\arguments{
\item{gene_of_int}{(char) gene of interest}

\item{system}{(char) system of interest}
}
\description{
\code{seqSimVis()} Calculates sequence similarity and plots dot plots based on the
three highest similarity values to visualize
}
\examples{
seqSimVis("BRCA1", "PHALY")

}
