% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyDButils.R
\name{SyDBgetIDforKey}
\alias{SyDBgetIDforKey}
\title{\code{SyDBgetIDforKey} fetch an ID from a systems database.}
\usage{
SyDBgetIDforKey(val, att, tbl, db, all = FALSE)
}
\arguments{
\item{val}{(character)  IDs matching value(s) "val" ...}

\item{att}{(character)  ... in any attribute column "att" ...}

\item{tbl}{(character)  ... of the table "tbl" ...}

\item{db}{(character)  in database "db".}

\item{all}{(logical)  if TRUE, then NA will be added to the returned
vector, one for each element in "val" that was
not found in "att". If FALSE (default), only
matching IDs will be returned, i.e. the length
of the input is not preserved.}
}
\value{
(character) a vector of QQIDs, or NA where
                    none were found
}
\description{
\code{SyDBgetIDforKey} documentation in progress ...
}
\examples{
\dontrun{
# get the IDs for the systems with code "PHALY" and "nosuch". Element 2 is NA.
SyDBgetIDforKey(c("PHALY", "nosuch"), "code","system", mySDB, all = TRUE)
}
}
