% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyDButils.R
\name{SyDBgetSysSymbols}
\alias{SyDBgetSysSymbols}
\title{\code{SyDBgetSysSymbols} fetch root-system IDs from a systems database.}
\usage{
SyDBgetSysSymbols(db, sys, treatNOT = "exclude")
}
\arguments{
\item{db}{(character)  name of the database}

\item{sys}{(character)  system codes or system IDs}

\item{treatNOT}{(character)  not implemented}
}
\value{
(list) a list of the same length as "sys", containing character
               vectors of HGNC symbols, or zero-length vectors if no
               symbols were found
}
\description{
\code{SyDBgetSysSymbols} documentation in progress ...
}
\examples{
\dontrun{
# all HGNC symbols of the PHALY system
SyDBgetSysSymbols(mySDB, "PHALY")

# all genes in the database:
unlist(SyDBgetSysSymbols(mySDB, SyDBgetRootSysIDs(mySDB)), use.names = FALSE)
}
}
