% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESA.R
\name{systemAnalysis}
\alias{systemAnalysis}
\title{\code{systemAnalysis()} Investigates genes from the system PHALY
and produces plots with gene network data}
\usage{
systemAnalysis(gene)
}
\arguments{
\item{sys}{Biological system}
}
\value{
a histogram for the node degree distribution of the PHALY gene network
and a network plot for a gene from the PHALY system
}
\description{
\code{systemAnalysis()} Investigates genes from the system PHALY
and produces plots with gene network data
}
\examples{
systemAnalysis("AMBRA1")

}
\author{
{Deus Bajaj} (aut)
}
