% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESA.R
\name{systemAnalysis}
\alias{systemAnalysis}
\title{\code{systemAnalysis()} Investigates genes from a biological system
and produces plots with gene network data}
\usage{
systemAnalysis(gene, sys)
}
\arguments{
\item{gene}{A gene from a Biological system}

\item{sys}{A Biological system}
}
\value{
a histogram for the node degree distribution of the system gene network
and a network plot for a gene from the biological system
}
\description{
\code{systemAnalysis()} Investigates genes from a biological system
and produces plots with gene network data
}
\examples{
\dontrun{
# Retrieve a histogram for the node degrees of the PHALY gene network and
# get the network plot of thr AMBRA1 gene from the PHALY system
systemAnalysis("AMBRA1", "PHALY")
}

}
\author{
{Deus Bajaj} (aut)
}
