% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesize.R
\name{visualizeInteractions}
\alias{visualizeInteractions}
\title{visualizeInteractions}
\usage{
visualizeInteractions(network, emap, ppi_ggi, gmap)
}
\arguments{
\item{network}{(dataframe) 3-column dataframe of physically-interacting
system components with their genetic interactions. The first two columns
denote the interacting pair; the third is the type of genetic interaction.}

\item{emap}{The dataframe obtained from \code{makeEMAP}.}

\item{ppi_ggi}{(dataframe) An optional 3-column dataframe to specify subset
of \code{mySys} for which both PPI and GGI data is available. The first two
columns denote the interacting pair; the third is the type of genetic
interaction.}

\item{gmap}{The dataframe obtained from \code{makeGMAP}.}
}
\description{
visualizeInteractions
}
\author{
Nada Elnour, \email{nada.elnour@mail.utoronto.ca}

A helper function of \code{hypothesize} that plots the hypothesis graph
}
